Use my take-screenshot module instead of the `canvas` module to generate image.
[See my demo of the screenshot module here](https://github.com/BillyZac/phantom-screenshot-demo).

So right now, the bot responds to mentions by taking a screenshot of the Flickr home page (or a portion of the Flickr page, anyways) and then posts that image to Twitter along with a non-sensical comment.

The bot does not @mention the original mentioner, to avoid violating Twitter's TOS while testing. I'll turn the @mentions back on when I feel more confident about the bot's behavior.